requirements.txt
asgiref==3.8.1
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
Markdown==3.7
mysqlclient==2.2.7
sqlparse==0.5.3
ujson==5.10.0
